# backend/docs/swagger/swagger.yaml
basePath: /
...
host: localhost:8088
info:
  contact: {}
  description: Bilingual CMS backend API for Impress (印迹). Supports content management,
    articles, pages, themes, media, and more.
  title: Impress CMS API
  version: "1.0"
paths:
...
  /admin/menus:
    get:
      description: Returns all menu groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              items:
                items:
                  type: object
                type: array
            type: object
      security:
      - BearerAuth: []
      summary: List menu groups
      tags:
      - Menus
    post:
      consumes:
      - application/json
      description: Create a new menu group
      parameters:
      - description: Menu group data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create menu group
      tags:
      - Menus
  /admin/menus/{id}:
    delete:
      description: Delete a menu group and its items
      parameters:
      - description: Menu group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete menu group
      tags:
      - Menus
    get:
      description: Returns a menu group with its items as a flat list
      parameters:
      - description: Menu group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get menu group
      tags:
      - Menus
    put:
      consumes:
      - application/json
      description: Update a menu group's name and slug
      parameters:
      - description: Menu group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated menu group data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update menu group
      tags:
      - Menus
  /admin/menus/{id}/items:
    post:
      consumes:
      - application/json
      description: Create a new menu item within a menu group
      parameters:
      - description: Menu group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Menu item data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create menu item
      tags:
      - Menus
  /admin/menus/{id}/items/{itemId}:
    delete:
      description: Delete a menu item by ID
      parameters:
      - description: Menu group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Menu item ID
        in: path
        name: itemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete menu item
      tags:
      - Menus
    put:
      consumes:
      - application/json
      description: Update a menu item's properties
      parameters:
      - description: Menu group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Menu item ID
        in: path
        name: itemId
        required: true
        type: integer
      - description: Updated menu item data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update menu item
      tags:
      - Menus
  /admin/menus/{id}/items/reorder:
    put:
      consumes:
      - application/json
      description: Set the display order of menu items in a group
      parameters:
      - description: Menu group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ordered item IDs
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reorder menu items
      tags:
      - Menus
  /admin/menus/{id}/primary:
    put:
      description: Set a menu group as the primary navigation menu
      parameters:
      - description: Menu group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set primary menu
      tags:
      - Menus
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}" for JWT authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"